# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для управления подписками пользователей. Позволяет создавать, получать, обновлять и удалять подписки, а также считать их общую стоимость.",
        "title": "Subscriptions Service API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/subscriptions": {
            "get": {
                "description": "Получает список всех подписок",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Получение списка подписок",
                "responses": {
                    "200": {
                        "description": "Список подписок",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.GetSubscriptionResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Создает новую запись о подписке в базе данных на основе переданных данных",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Добавление новой подписки",
                "parameters": [
                    {
                        "description": "Данные для создания подписки",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateSubscriptionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "ID созданной подписки",
                        "schema": {
                            "$ref": "#/definitions/dto.AddSubscriptionResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный формат JSON или некорректные данные",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/subscriptions/cost": {
            "get": {
                "description": "Возращает суммарную стоимость всех подписок",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Получение общей стоимости подписок",
                "parameters": [
                    {
                        "type": "string",
                        "example": "123e4567-e89b-12d3-a456-426614174000",
                        "description": "ID пользователя (UUID)",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "Netflix",
                        "description": "Название сервиса",
                        "name": "service",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "01-2026",
                        "description": "Дата начала (MM-YYYY)",
                        "name": "from_date",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "07-2026",
                        "description": "Дата окончания (MM-YYYY)",
                        "name": "end_date",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Суммарная стоимость подписок",
                        "schema": {
                            "$ref": "#/definitions/dto.TotalCostResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный формат JSON или некорректные данные",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/subscriptions/{id}": {
            "get": {
                "description": "Возвращает информацию о подписке по ее идентификатору",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Получение подписки по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "example": 123,
                        "description": "ID подписки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о подписке",
                        "schema": {
                            "$ref": "#/definitions/dto.GetSubscriptionResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный формат ID",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет подписку по идентификатору",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Удаление подписки",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "example": 123,
                        "description": "ID подписки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Подписка успешно удалена"
                    },
                    "400": {
                        "description": "Неверный формат ID",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Обновляет данные существующей подписки",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Обновление подписки",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "example": 123,
                        "description": "ID подписки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные для обновления подписки",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateSubscriptionRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Подписка успешно обновлена"
                    },
                    "400": {
                        "description": "Неверный формат JSON или некорректные данные",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.AddSubscriptionResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 123
                }
            }
        },
        "dto.CreateSubscriptionRequest": {
            "type": "object",
            "properties": {
                "end_date": {
                    "type": "string",
                    "example": "07-2025"
                },
                "price": {
                    "type": "integer",
                    "example": 399
                },
                "service_name": {
                    "type": "string",
                    "example": "Yandex Plus"
                },
                "start_date": {
                    "type": "string",
                    "example": "01-2026"
                },
                "user_id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                }
            }
        },
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Internal server error"
                }
            }
        },
        "dto.GetSubscriptionResponse": {
            "type": "object",
            "properties": {
                "end_date": {
                    "type": "string",
                    "example": "07-2026"
                },
                "id": {
                    "type": "integer",
                    "example": 123
                },
                "price": {
                    "type": "integer",
                    "example": 399
                },
                "service_name": {
                    "type": "string",
                    "example": "Yandex Plus"
                },
                "start_date": {
                    "type": "string",
                    "example": "01-2026"
                },
                "user_id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                }
            }
        },
        "dto.TotalCostResponse": {
            "type": "object",
            "properties": {
                "total_cost": {
                    "type": "integer",
                    "example": 1799
                }
            }
        },
        "dto.UpdateSubscriptionRequest": {
            "type": "object",
            "properties": {
                "end_date": {
                    "type": "string",
                    "example": "07-2026"
                },
                "price": {
                    "type": "integer",
                    "example": 399
                }
            }
        }
    }
}